move some utils to components
