+++ src/components/placeSymbol.js
