+++ src/components/placeSymbol.test.js
