+++ src/components/startGame.js
